RC #191 - Adding stories for bibliography components; Fixing a bug in CitationStyles hook
