+++ packages/semantic-ui/src/components/BibliographyList.js
@@ -23,7 +23,7 @@ import SortSelector from './SortSelector';
 import { SORT_DESCENDING } from '../constants/Sort';
 import StyleSelector from './StyleSelector';
 import Toaster from './Toaster';
-import useList from './List';
+import useList, { Props as ListProps } from './List';
 import ZoteroTranslateContext from '../context/ZoteroTranslateContext';
 import './BibliographyList.css';
 
@@ -110,13 +110,32 @@ const BibliographyListComponent: ComponentType<any> = useList((props: ComponentP
   </List>
 ));
 
-type Props = {
+type Props = ListProps & {
+  /**
+   * The array of bibliography items.
+   */
   items: Array<Item>,
+
+  /**
+   * Callback fired when the passed item is deleted. This function should return a Promise.
+   */
   onDelete: (item: Item) => Promise<any>,
+
+  /**
+   * Callback fired when the passed item is saved. This function should return a Promise.
+   */
   onSave: (item: Item) => Promise<any>,
+
+  /**
+   * URL of the Zotero translation server.
+   */
   translateUrl: string
 };
 
+/**
+ * This component renders a list of bibliography items in Zbib format. A Zotero translation server is required in
+ * order to format the data and display the conditional form.
+ */
 const BibliographyList = (props: Props) => {
   const [locale, setLocale] = useState();
   const [showToaster, setShowToaster] = useState(false);
