+++ packages/semantic-ui/src/index.js
@@ -10,7 +10,9 @@ export { default as AccordionSelector } from './components/AccordionSelector';
 export { default as ArrowButtons } from './components/ArrowButtons';
 export { default as AssociatedDropdown } from './components/AssociatedDropdown';
 export { default as AudioPlayer } from './components/AudioPlayer';
+export { default as BibliographyForm } from './components/BibliographyForm';
 export { default as BibliographyList } from './components/BibliographyList';
+export { default as BibliographySearchInput } from './components/BibliographySearchInput';
 export { default as BooleanIcon } from './components/BooleanIcon';
 export { default as CancelButton } from './components/CancelButton';
 export { default as ColorButton } from './components/ColorButton';
@@ -85,6 +87,7 @@ export { default as Section } from './components/Section';
 export { default as Selectize } from './components/Selectize';
 export { default as SelectizeHeader } from './components/SelectizeHeader';
 export { default as SelectizeImageHeader } from './components/SelectizeImageHeader';
+export { default as StyleSelector } from './components/StyleSelector';
 export { default as TabbedModal } from './components/TabbedModal';
 export { default as TabsMenu } from './components/TabsMenu';
 export { default as TagsList } from './components/TagsList';
